% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineaHorizontal.R
\name{lineasignif}
\alias{lineasignif}
\title{Barra de significación horizontal
Inserta barra de significación horizontal.}
\usage{
lineasignif(
  data,
  linea_inicio,
  linea_fin,
  linea_pos,
  linea_izq_pos = NULL,
  linea_der_pos = NULL,
  linea_izq_long = NULL,
  linea_der_long = NULL,
  texto,
  texto_ajusta_pos_v = 0,
  texto_size = 3.5,
  texto_familia = "sans",
  texto_estilo = "plain",
  texto_color = "black",
  linea_color = "black",
  linea_grosor = 1,
  linea_estilo = "solid",
  facet_column = NULL,
  facet_value = NULL
)
}
\arguments{
\item{data}{Nombre del dataframe.}

\item{linea_inicio}{Posición del eje x inicia la línea.}

\item{linea_fin}{Posición del eje x termina la línea.}

\item{linea_pos}{Posición de la línea en el eje y.}

\item{linea_izq_pos}{Posición donde termina la línea vertical extrema de la izquierda. Se puede remplazar con linea_izq_long}

\item{linea_der_pos}{Posición donde termina la línea vertical extrema de la derecha. Se puede remplazar con linea_der_long}

\item{linea_izq_long}{Longitud de la línea vertical extrema de la izquierda. Se puede remplazar con linea_izq_pos.}

\item{linea_der_long}{Longitud de la línea vertical extrema de la derecha. . Se puede remplazar con linea_der_pos.}

\item{texto}{Texto que se va a mostrar sobre la línea Ej. “P<0.005”}

\item{texto_ajusta_pos_v}{Ajusta la posición del texto con respecto a la línea, se puede subir o bajar. Se usa la escala de medida del eje y.}

\item{texto_size}{Ajusta el tamaño del texto}

\item{texto_familia}{Ajusta la familia del texto}

\item{texto_estilo}{Ajusta el estilo del texto (Negritas, cursiva, etc.)}

\item{texto_color}{Ajusta el color del texto}

\item{linea_color}{Cambia el color de las líneas}

\item{linea_grosor}{Cambia el grosor de las líneas}

\item{linea_estilo}{Cambia el estilo de las líneas (Continua, punteada, etc.)}

\item{facet_column}{Nombre de la columna del DataFrame se está utilizando para facetado en el gráfico ggplot2. Por ejemplo, si tu DataFrame tiene una columna 'grupo' que define las facetas, deberías establecer facet_column = "grupo".}

\item{facet_value}{Este parámetro se utiliza en combinación con facet_column para especificar en qué faceta particular quieres que aparezcan las líneas y el texto. Por ejemplo, si la columna 'grupo' tiene valores 'A', 'B' y 'C', y deseas dibujar en la faceta correspondiente a 'B', debes establecer facet_value = "B".}
}
\value{
LineaHorizontal
}
\description{
Barra de significación horizontal
Inserta barra de significación horizontal.
}
\examples{
 data <- data.frame(
manufacturer = c('Manufacturer A', 'Manufacturer B',
                'Manufacturer C', 'Manufacturer D', 'Manufacturer E'),
cyl = c(4, 7, 6, 5, 8)  # Valores específicos en lugar de aleatorios
)
# Creating the plot
p <- ggplot(data, aes(x = manufacturer, y = cyl)) +
 geom_col(fill = '#112446') +
 theme_minimal() +
 xlab('Manufacturer') +
 ylab('Count of Cars') +
 ggtitle('Car Counts by Manufacturer') +
 theme(axis.text.x = element_text(angle = 45, hjust = 1))+
 ylim(0, 12) #Es para que el texto no se recorte+
coord_flip()

print(p)

#Crear la linea de significación
ln <- lineasignif(data,
                 linea_pos = 10,
                 linea_inicio = 1,
                 linea_fin = 5,
                 texto = "P= 0.005",
                 linea_izq_long =  5,
                 linea_der_pos =  8.3,
                 linea_grosor = 3)

p+ln


}
